requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
httpx>=0.24.0
beautifulsoup4>=4.9.3
python-dotenv>=0.19.0
jinja2>=2.11.3
aiofiles>=0.6.0 